+++ docs/changelog.rst
@@ -18,6 +18,12 @@ Change Log
 Version 1.x
 ===========
 
+.. version:: 1.2
+   :date: 2024-09-19
+
+   - Skip document with ``:nocomment:`` docinfos (:pull:`5`)
+   - Allow user specifying Isso thread ID via ``:isso-id:`` docinfo (:pull:`5`)
+
 .. version:: 1.1
    :date: 2024-05-01